requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.25
certifi==2017.11.5
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.9
django-activelink==0.4
django-celery==3.2.1
django-configurations==2.0
django-debug-toolbar==1.8
django-rest-swagger==2.1.2
djangorestframework==3.7.7
drf-openapi==1.3.0
drfdocs==0.0.11
ephem==3.7.6.0
idna==2.6
ipython==1.2.1
itypes==1.1.0
Jinja2==2.10
kombu==3.0.37
MarkupSafe==1.0
MySQL-python==1.2.5
openapi-codec==1.3.2
pytz==2017.3
requests==2.18.4
simplejson==3.13.2
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.22
uWSGI==2.0.15
vine==1.1.4
